pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "trading-algorithm"
version = "0.1.0"
description = "A backtesting and live trading system for TECL/VIX strategies"
authors = [{name = "Nitin Rao"}]
readme = "README.md"
requires-python = ">=3.9"
license = {file = "LICENSE"}
keywords = ["trading", "backtesting", "finance", "tecl", "vix"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "alpaca-py>=0.8.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "yfinance>=0.2.0",
    "finnhub-python>=2.4.0",
    "alpha-vantage>=2.3.0",
    "schedule>=1.2.0",
    "pytz>=2023.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.0.0",
]
live = [
    "alpaca-py>=0.8.0",
    "python-dotenv>=1.0.0",
]
analysis = [
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "jupyter>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/nitinrao/trading-algorithm"
Repository = "https://github.com/nitinrao/trading-algorithm"
Issues = "https://github.com/nitinrao/trading-algorithm/issues"

[project.scripts]
trading-backtest = "trading_algorithm.backtesting:cli_main"
trading-live = "trading_algorithm.live_trader:main"
trading-monitor = "trading_algorithm.scheduler:start_scheduler"

[tool.hatch.build.targets.wheel]
packages = ["trading_algorithm"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
target-version = "py39"
line-length = 100
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=trading_algorithm --cov-report=term-missing --cov-report=html"